Added acceptance test for validate ordering of answers
