+++ app/views/questions/_uno_answer.html.slim
@@ -1,4 +1,4 @@
-div id="answer-#{answer.id}"
+div.answer id="answer-#{answer.id}"
   p.best-answer
     = render partial: 'questions/best_label', locals: {answer: answer}
   p.answer-body
