+++ spec/acceptance/question_features_spec.rb
@@ -121,73 +121,3 @@ feature 'Delete question' do
   end
 end
 
-feature 'Edit question' do
-  given(:owner){ create(:user)}
-  given(:user2){ create(:user)}
-  given(:question){ create(:question, user: owner) }
-
-  scenario 'Owner can edit question', js: true do
-    sign_in(owner)
-    visit question_path(question)
-    within (".question") do
-      click_on 'Edit'
-      fill_in 'Title', with: 'New title question'
-      fill_in 'Body', with: 'New body question'
-      click_on 'Save'
-
-      expect(page).to have_content 'New title question'
-      expect(page).to have_content 'New body question'
-    end
-  end
-
-  scenario 'Not owner can\'t edit question' do
-    sign_in(user2)
-    visit question_path(question)
-    within('.question') do
-      expect(page).to_not have_content 'Edit'
-     end
-  end
-
-  scenario 'Non-authenticated user can\'t edit question' do
-    visit question_path(question)
-    within('.question') do
-      expect(page).to_not have_content 'Edit'
-    end
-  end
-end
-
-feature 'Select best answer to the question' do
-  given(:user1){ create(:user)}
-  given(:user2){ create(:user)}
-  given(:question){ create(:question, user: user1) }
-  given!(:answer){create(:answer, question: question, user: user2)}
-  given!(:answer2){create(:answer, question: question, user: user2)}
-
-  scenario 'Question\'s owner can select one of answers as Best Answer', js: true do
-    sign_in(user1)
-    visit question_path(question)
-    within("#answer-#{answer.id}") do
-      click_on 'Set best'
-    end
-    within("#answer-#{answer2.id}") do
-      click_on 'Set best'
-    end
-    within("#answer-#{answer.id}") do
-      expect(page).to_not have_content 'Best answer'
-    end
-    within("#answer-#{answer2.id}") do
-      expect(page).to have_content 'Best answer'
-    end
-  end
-
-  scenario 'Not owner can\'t select answers as Best Answer' do
-    sign_in(user2)
-    visit question_path(question)
-    expect(page).to_not have_content 'Set best'
-  end
-
-  scenario 'Non-authenticated user can\'t select answers as Best Answer' do
-    visit question_path(question)
-    expect(page).to_not have_content 'Set best'
-  end
-end
